+++ static/components/utils.js
@@ -14,7 +14,9 @@ export function pushOptions(el, options) {
 
 /** Show or hide an element based on the current view */
 const hash = location.hash.slice(1);
-export const initialViews = new Set(hash ? hash.split(",") : []);
+export const initialViews = new Set(
+  hash ? hash.split(",").map(decodeURIComponent) : [],
+);
 export function view(element) {
   const { schema } = element;